app/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
python-dotenv
openai